Remaking music

+++ src/GUI/roundedBackplate.cpp
@@ -8,12 +8,12 @@
 
 using namespace GUI;
 
-//
+// Class of backplates (smoothed rects)
 Backplate::Backplate(float _centerX, float _centerY, float _width, float _height,
     Uint8 _rad, Uint8 _bor, SDL_Color _frontColor, SDL_Color _backColor)
 : frontColor(_frontColor), backColor(_backColor), rad(_rad), bor(_bor) {
-    updatePlate({SCREEN_WIDTH * (_centerX - _width/2), SCREEN_HEIGHT * (_centerY - _height/2),
-        SCREEN_WIDTH * _width, SCREEN_HEIGHT * _height});
+    updatePlate({int(SCREEN_WIDTH * (_centerX - _width/2)), int(SCREEN_HEIGHT * (_centerY - _height/2)),
+        int(SCREEN_WIDTH * _width), int(SCREEN_HEIGHT * _height)});
 }
 
 Backplate::Backplate(const Uint8 _rad, const Uint8 _bor, const SDL_Color _frontColor, const SDL_Color _backColor)
@@ -21,12 +21,10 @@ Backplate::Backplate(const Uint8 _rad, const Uint8 _bor, const SDL_Color _frontC
     rect = {0, 0, 0, 0};
 }
 
-//
 Backplate::~Backplate() {
     SDL_DestroyTexture(texture);
 }
 
-//
 void Backplate::updatePlate(const SDL_Rect _rect) {
     // Checking, if need clear previous texture
     if (texture) {
